+++ src/core/zowe/core_for_zowe_sdk/config_file.py
@@ -10,46 +10,32 @@ SPDX-License-Identifier: EPL-2.0
 Copyright Contributors to the Zowe Project.
 """
 
-import base64
+import json
 import os.path
 import re
-import json
-import requests
-import sys
 import warnings
+from copy import deepcopy
 from dataclasses import dataclass, field
-from typing import Optional, NamedTuple
+from typing import NamedTuple, Optional
 
 import commentjson
+import requests
 
-from .constants import constants
-from .custom_warnings import (
-    ProfileNotFoundWarning,
-    ProfileParsingWarning,
-    SecurePropsNotFoundWarning,
-)
-from .exceptions import ProfileNotFound, SecureProfileLoadFailed
-from .profile_constants import (
-    GLOBAL_CONFIG_NAME,
-    TEAM_CONFIG,
-    USER_CONFIG,
-)
-
-HAS_KEYRING = True
-try:
-    import keyring
-except ImportError:
-    HAS_KEYRING = False
+from .credential_manager import CredentialManager
+from .custom_warnings import ProfileNotFoundWarning, ProfileParsingWarning
+from .exceptions import ProfileNotFound
+from .profile_constants import GLOBAL_CONFIG_NAME, TEAM_CONFIG, USER_CONFIG
+from .validators import validate_config_json
 
 HOME = os.path.expanduser("~")
-GLOBAl_CONFIG_LOCATION = os.path.join(HOME, ".zowe")
-GLOBAL_CONFIG_PATH = os.path.join(
-    GLOBAl_CONFIG_LOCATION, f"{GLOBAL_CONFIG_NAME}.config.json"
-)
+GLOBAL_CONFIG_LOCATION = os.path.join(HOME, ".zowe")
+GLOBAL_CONFIG_PATH = os.path.join(GLOBAL_CONFIG_LOCATION, f"{GLOBAL_CONFIG_NAME}.config.json")
 CURRENT_DIR = os.getcwd()
 
 # Profile datatype is used by ConfigFile to return Profile Data along with
 # metadata such as profile_name and secure_props_not_found
+
+
 class Profile(NamedTuple):
     data: dict = {}
     name: str = ""
@@ -80,8 +66,9 @@ class ConfigFile:
     _location: Optional[str] = None
     profiles: Optional[dict] = None
     defaults: Optional[dict] = None
-    secure_props: Optional[dict] = None
     schema_property: Optional[dict] = None
+    secure_props: Optional[dict] = None
+    jsonc: Optional[dict] = None
     _missing_secure_props: list = field(default_factory=list)
 
     @property
@@ -107,7 +94,7 @@ class ConfigFile:
 
     @property
     def schema_path(self) -> Optional[str]:
-        self.schema_property
+        return self.schema_property
 
     @location.setter
     def location(self, dirname: str) -> None:
@@ -116,25 +103,56 @@ class ConfigFile:
         else:
             raise FileNotFoundError(f"given path {dirname} is not valid")
 
-    def init_from_file(self) -> None:
+    def init_from_file(
+        self,
+        validate_schema: Optional[bool] = True,
+    ) -> None:
         """
         Initializes the class variable after
         setting filepath (or if not set, autodiscover the file)
         """
         if self.filepath is None:
-            self.autodiscover_config_dir()
+            try:
+                self.autodiscover_config_dir()
+            except FileNotFoundError:
+                pass
+
+        if self.filepath is None or not os.path.isfile(self.filepath):
+            warnings.warn(f"Config file does not exist at {self.filepath}")
+            return
 
         with open(self.filepath, encoding="UTF-8", mode="r") as fileobj:
             profile_jsonc = commentjson.load(fileobj)
 
         self.profiles = profile_jsonc.get("profiles", {})
-        self.defaults = profile_jsonc.get("defaults", {})
         self.schema_property = profile_jsonc.get("$schema", None)
+        self.defaults = profile_jsonc.get("defaults", {})
+        self.jsonc = profile_jsonc
 
-        # loading secure props is done in load_profile_properties
-        # since we want to try loading secure properties only when
-        # we know that the profile has saved properties
-        # self.load_secure_props()
+        if self.schema_property and validate_schema:
+            self.validate_schema()
+
+        CredentialManager.load_secure_props()
+        self.__load_secure_properties()
+
+    def validate_schema(self) -> None:
+        """
+        Get the $schema_property from the config and load the schema
+
+        Returns
+        -------
+        file_path to the $schema property
+        """
+
+        path_schema_json = None
+
+        path_schema_json = self.schema_path
+        if path_schema_json is None:  # check if the $schema property is not defined
+            warnings.warn(f"$schema property could not found")
+
+        # validate the $schema property
+        if path_schema_json:
+            validate_config_json(self.jsonc, path_schema_json, cwd = self.location)
 
     def schema_list(
         self,
@@ -142,11 +160,11 @@ class ConfigFile:
         """
         Loads the schema properties
         in a sorted order according to the priority
-        
+
         Returns
         -------
         Dictionary
-        
+
             Returns the profile properties from schema (prop: value)
         """
 
@@ -157,23 +175,23 @@ class ConfigFile:
         if schema.startswith("https://") or schema.startswith("http://"):
             schema_json = requests.get(schema).json()
 
+        elif os.path.isfile(schema) or schema.startswith("file://"):
+            with open(schema.replace("file://", "")) as f:
+                schema_json = json.load(f)
+
         elif not os.path.isabs(schema):
             schema = os.path.join(self.location, schema)
-            with open(schema) as f:
-                schema_json = json.load(f)
-        
-        elif os.path.isfile(schema):
             with open(schema) as f:
                 schema_json = json.load(f)
         else:
             return []
 
-        profile_props:dict = {}
+        profile_props: dict = {}
         schema_json = dict(schema_json)
-        
-        for props in schema_json['properties']['profiles']['patternProperties']["^\\S*$"]["allOf"]:
+
+        for props in schema_json["properties"]["profiles"]["patternProperties"]["^\\S*$"]["allOf"]:
             props = props["then"]
-            
+
             while "properties" in props:
                 props = props.pop("properties")
                 profile_props = props
@@ -184,6 +202,7 @@ class ConfigFile:
         self,
         profile_name: Optional[str] = None,
         profile_type: Optional[str] = None,
+        validate_schema: Optional[bool] = True,
     ) -> Profile:
         """
         Load given profile including secure properties and excluding values from base profile
@@ -193,7 +212,7 @@ class ConfigFile:
             Returns a namedtuple called Profile
         """
         if self.profiles is None:
-            self.init_from_file()
+            self.init_from_file(validate_schema)
 
         if profile_name is None and profile_type is None:
             raise ProfileNotFound(
@@ -202,10 +221,7 @@ class ConfigFile:
             )
 
         if profile_name is None:
-            profile_name = self.get_profilename_from_profiletype(
-                profile_type=profile_type
-            )
-
+            profile_name = self.get_profilename_from_profiletype(profile_type=profile_type)
         props: dict = self.load_profile_properties(profile_name=profile_name)
 
         return Profile(props, profile_name, self._missing_secure_props)
@@ -262,7 +278,7 @@ class ConfigFile:
             return profilename
 
         # iterate through the profiles and check if profile is found
-        for (key, value) in self.profiles.items():
+        for key, value in self.profiles.items():
             try:
                 temp_profile_type = value["type"]
                 if profile_type == temp_profile_type:
@@ -278,7 +294,7 @@ class ConfigFile:
             profile_name=profile_type,
             error_msg=f"No profile with matching profile_type '{profile_type}' found",
         )
-        
+
     def find_profile(self, path: str, profiles: dict):
         """
         Find a profile at a specified location from within a set of nested profiles
@@ -309,7 +325,6 @@ class ConfigFile:
         Load exact profile properties (without prepopulated fields from base profile)
         from the profile dict and populate fields from the secure credentials storage
         """
-
         props = {}
         lst = profile_name.split(".")
         secure_fields: list = []
@@ -318,78 +333,166 @@ class ConfigFile:
             profile_name = ".".join(lst)
             profile = self.find_profile(profile_name, self.profiles)
             if profile is not None:
-                props = { **profile.get("properties", {}), **props }
+                props = {**profile.get("properties", {}), **props}
                 secure_fields.extend(profile.get("secure", []))
             else:
-                warnings.warn(
-                        f"Profile {profile_name} not found",
-                        ProfileNotFoundWarning
-                        )
+                warnings.warn(f"Profile {profile_name} not found", ProfileNotFoundWarning)
             lst.pop()
-
-
-        # load secure props only if there are secure fields
-        if secure_fields:
-            self.load_secure_props()
-
-            # load properties with key as profile.{profile_name}.properties.{*}
-            for (key, value) in self.secure_props.items():
-                if re.match(
-                    "profiles\\." + profile_name + "\\.properties\\.[a-z]+", key
-                ):
-                    property_name = key.split(".")[3]
-                    if property_name in secure_fields:
-                        props[property_name] = value
-                        secure_fields.remove(property_name)
-
-            # if len(secure_fields) > 0:
-            #     self._missing_secure_props.extend(secure_fields)
 
         return props
 
-    def load_secure_props(self) -> None:
+    def __load_secure_properties(self):
         """
-        load secure_props stored for the given config file
-        Returns
-        -------
-        None
+        Inject secure properties that have been loaded from the vault into the profiles object.
+        """
+        secure_props = CredentialManager.secure_props.get(self.filepath, {})
+        for key, value in secure_props.items():
+            segments = [name for i, name in enumerate(key.split(".")) if i % 2 == 1]
+            profiles_obj = self.profiles
+            property_name = segments.pop()
+            for i, profile_name in enumerate(segments):
+                if profile_name in profiles_obj:
+                    profiles_obj = profiles_obj[profile_name]
+                    if i == len(segments) - 1:
+                        profiles_obj.setdefault("properties", {})
+                        profiles_obj["properties"][property_name] = value
+                else:
+                    break
 
-        if keyring is not initialized, set empty value
+    def __extract_secure_properties(self, profiles_obj, json_path="profiles"):
         """
-        if not HAS_KEYRING:
-            self.secure_props = {}
+        Extract secure properties from the profiles object so they can be saved to the vault.
+        """
+        secure_props = {}
+        for key, value in profiles_obj.items():
+            for property_name in value.get("secure", []):
+                if property_name in value.get("properties", {}):
+                    secure_props[f"{json_path}.{key}.properties.{property_name}"] = value["properties"].pop(
+                        property_name
+                    )
+            if value.get("profiles"):
+                secure_props.update(self.__extract_secure_properties(value["profiles"], f"{json_path}.{key}.profiles"))
+        return secure_props
+
+    def __set_or_create_nested_profile(self, profile_name, profile_data):
+        """
+        Set or create a nested profile.
+        """
+        path = self.get_profile_path_from_name(profile_name)
+        keys = path.split(".")[1:]
+        nested_profiles = self.profiles
+        for key in keys:
+            nested_profiles = nested_profiles.setdefault(key, {})
+        nested_profiles.update(profile_data)
+
+    def __is_secure(self, json_path: str, property_name: str) -> bool:
+        """
+        Check whether the given JSON path corresponds to a secure property.
+
+        Parameters:
+            json_path (str): The JSON path of the property to check.
+            property_name (str): The name of the property to check.
+
+        Returns:
+            bool: True if the property should be stored securely, False otherwise.
+        """
+
+        profile = self.find_profile(json_path, self.profiles)
+        if profile and profile.get("secure"):
+            return property_name in profile["secure"]
+        return False
+
+    def set_property(self, json_path, value, secure=None) -> None:
+        """
+        Set a property in the profile, storing it securely if necessary.
+
+        Parameters:
+            json_path (str): The JSON path of the property to set.
+            value (str): The value to be set for the property.
+            profile_name (str): The name of the profile to set the property in.
+            secure (bool): If True, the property will be stored securely. Default is None.
+        """
+        if self.profiles is None:
+            self.init_from_file()
+
+        # Checking whether the property should be stored securely or in plain text
+        property_name = json_path.split(".")[-1]
+        profile_name = self.get_profile_name_from_path(json_path)
+        # check if the property is already secure
+        is_property_secure = self.__is_secure(profile_name, property_name)
+        is_secure = secure if secure is not None else is_property_secure
+
+        current_profile = self.find_profile(profile_name, self.profiles) or {}
+        current_properties = current_profile.setdefault("properties", {})
+        current_secure = current_profile.setdefault("secure", [])
+        current_properties[property_name] = value
+        if is_secure and not is_property_secure:
+            current_secure.append(property_name)
+        elif not is_secure and is_property_secure:
+            current_secure.remove(property_name)
+
+        current_profile["properties"] = current_properties
+        current_profile["secure"] = current_secure
+        self.__set_or_create_nested_profile(profile_name, current_profile)
+
+    def set_profile(self, profile_path: str, profile_data: dict) -> None:
+        """
+        Set a profile in the config file.
+
+        Parameters:
+            profile_path (str): The path of the profile to be set. eg: profiles.zosmf
+            profile_data (dict): The data to be set for the profile.
+        """
+        if self.profiles is None:
+            self.init_from_file()
+        profile_name = self.get_profile_name_from_path(profile_path)
+        if "secure" in profile_data:
+            # Checking if the profile has a 'secure' field with values
+            secure_fields = profile_data["secure"]
+            current_profile = self.find_profile(profile_name, self.profiles) or {}
+            existing_secure_fields = current_profile.get("secure", [])
+            new_secure_fields = [field for field in secure_fields if field not in existing_secure_fields]
+
+            # Updating the 'secure' field of the profile with the combined list of secure fields
+            profile_data["secure"] = existing_secure_fields + new_secure_fields
+            # If a field is provided in the 'secure' list and its value exists in 'profile_data', remove it
+            profile_data["properties"] = {
+                **current_profile.get("properties", {}),
+                **profile_data.get("properties", {}),
+            }
+        self.__set_or_create_nested_profile(profile_name, profile_data)
+
+    def save(self, update_secure_props=True):
+        """
+        Save the config file to disk. and secure props to vault
+        parameters:
+            secure_props (bool): If True, the secure properties will be stored in the vault. Default is True.
+        Returns:
+            None
+        """
+        # Updating the config file with any changes
+        if not any(self.profiles.values()):
             return
 
-        try:
-            service_name = constants["ZoweServiceName"]
+        profiles_temp = deepcopy(self.profiles)
+        secure_props = self.__extract_secure_properties(profiles_temp)
+        CredentialManager.secure_props[self.filepath] = secure_props
+        with open(self.filepath, "w") as file:
+            self.jsonc["profiles"] = profiles_temp
+            commentjson.dump(self.jsonc, file, indent=4)
+        if update_secure_props:
+            CredentialManager.save_secure_props()
 
-            if sys.platform == "win32":
-                service_name += "/" + constants["ZoweAccountName"]
+    def get_profile_name_from_path(self, path: str) -> str:
+        """
+        Get the name of the profile from the given path.
+        """
+        segments = path.split(".")
+        profile_name = ".".join(segments[i] for i in range(1, len(segments), 2) if segments[i - 1] != "properties")
+        return profile_name
 
-            secret_value = keyring.get_password(
-                service_name, constants["ZoweAccountName"]
-            )
-
-        except Exception as exc:
-            raise SecureProfileLoadFailed(
-                constants["ZoweServiceName"], error_msg=str(exc)
-            ) from exc
-
-        secure_config: str
-        if sys.platform == "win32":
-            secure_config = secret_value.encode("utf-16")
-        else:
-            secure_config = secret_value
-
-        secure_config_json = commentjson.loads(base64.b64decode(secure_config).decode())
-
-        # look for credentials stored for currently loaded config
-        try:
-            self.secure_props = secure_config_json.get(self.filepath, {})
-        except KeyError as exc:
-            error_msg = str(exc)
-            warnings.warn(
-                f"No credentials found for loaded config file '{self.filepath}'"
-                f" with error '{error_msg}'",
-                SecurePropsNotFoundWarning,
-            )
+    def get_profile_path_from_name(self, short_path: str) -> str:
+        """
+        Get the path of the profile from the given name.
+        """
+        return re.sub(r"(^|\.)", r"\1profiles.", short_path)