+++ src/core/zowe/core_for_zowe_sdk/session.py
@@ -45,7 +45,7 @@ class Session:
         if props.get("host") is not None:
             self.session: ISession = ISession(host=props.get("host"))
         else:
-            raise "Host must be supplied"
+            raise Exception("Host must be supplied")
 
         # determine authentication type
         if props.get("user") is not None and props.get("password") is not None:
@@ -61,15 +61,13 @@ class Session:
             self.session.tokenValue = props.get("tokenValue")
             self.session.type = session_constants.AUTH_TYPE_BEARER
         else:
-            raise "An authentication method must be supplied"
+            raise Exception("An authentication method must be supplied")
 
         # set additional parameters
         self.session.basePath = props.get("basePath")
         self.session.port = props.get("port", self.session.port)
         self.session.protocol = props.get("protocol", self.session.protocol)
-        self.session.rejectUnauthorized = props.get(
-            "rejectUnauthorized", self.session.rejectUnauthorized
-        )
+        self.session.rejectUnauthorized = props.get("rejectUnauthorized", self.session.rejectUnauthorized)
 
     def load(self) -> ISession:
         return self.session