+++ src/core/zowe/core_for_zowe_sdk/zosmf_profile.py
@@ -13,15 +13,16 @@ Copyright Contributors to the Zowe Project.
 import base64
 import os.path
 import sys
+
 import yaml
 
+from .connection import ApiConnection
 from .constants import constants
 from .exceptions import SecureProfileLoadFailed
-from .connection import ApiConnection
 
 HAS_KEYRING = True
 try:
-    import keyring
+    from zowe.secrets_for_zowe_sdk import keyring
 except ImportError:
     HAS_KEYRING = False
 
@@ -32,8 +33,8 @@ class ZosmfProfile:
 
     Description
     -----------
-    This class is only used when there is already a Zowe z/OSMF profile created 
-    and the user opted to use the profile instead of passing the credentials directly 
+    This class is only used when there is already a Zowe z/OSMF profile created
+    and the user opted to use the profile instead of passing the credentials directly
     in the object constructor.
 
     Attributes
@@ -67,9 +68,7 @@ class ZosmfProfile:
         zosmf_connection
             z/OSMF connection object
         """
-        profile_file = os.path.join(
-            self.profiles_dir, "{}.yaml".format(self.profile_name)
-        )
+        profile_file = os.path.join(self.profiles_dir, "{}.yaml".format(self.profile_name))
 
         with open(profile_file, "r") as fileobj:
             profile_yaml = yaml.safe_load(fileobj)
@@ -80,30 +79,26 @@ class ZosmfProfile:
 
         zosmf_user = profile_yaml["user"]
         zosmf_password = profile_yaml["password"]
-        if zosmf_user.startswith(
+        if zosmf_user.startswith(constants["SecureValuePrefix"]) and zosmf_password.startswith(
             constants["SecureValuePrefix"]
-        ) and zosmf_password.startswith(constants["SecureValuePrefix"]):
+        ):
             zosmf_user, zosmf_password = self.__load_secure_credentials()
 
         zosmf_ssl_verification = True
         if "rejectUnauthorized" in profile_yaml:
             zosmf_ssl_verification = profile_yaml["rejectUnauthorized"]
 
-        return ApiConnection(
-            zosmf_host, zosmf_user, zosmf_password, zosmf_ssl_verification
-        )
+        return ApiConnection(zosmf_host, zosmf_user, zosmf_password, zosmf_ssl_verification)
 
     def __get_secure_value(self, name):
         service_name = constants["ZoweCredentialKey"]
         account_name = "zosmf_{}_{}".format(self.profile_name, name)
 
-        if sys.platform == "win32":
-            service_name += "/" + account_name
-
         secret_value = keyring.get_password(service_name, account_name)
 
-        if sys.platform == "win32":
-            secret_value = secret_value.encode("utf-16")
+        # Handle the case when secret_value is None
+        if secret_value is None:
+            secret_value = ""
 
         secret_value = base64.b64decode(secret_value).decode().strip('"')
 
@@ -112,9 +107,7 @@ class ZosmfProfile:
     def __load_secure_credentials(self):
         """Load secure credentials for a z/OSMF profile."""
         if not HAS_KEYRING:
-            raise SecureProfileLoadFailed(
-                self.profile_name, "Keyring module not installed"
-            )
+            raise SecureProfileLoadFailed(self.profile_name, "Keyring module not installed")
 
         try:
             zosmf_user = self.__get_secure_value("user")
@@ -123,37 +116,3 @@ class ZosmfProfile:
             raise SecureProfileLoadFailed(self.profile_name, e)
         else:
             return (zosmf_user, zosmf_password)
-
-
-if HAS_KEYRING and sys.platform.startswith("linux"):
-    from contextlib import closing
-    from keyring.backends import SecretService
-
-    class KeyringBackend(SecretService.Keyring):
-        """
-        Class used to handle secured profiles.
-
-        Methods
-        -------
-        get_password(service, username)
-            Get the decoded password
-        """
-
-        def __get_password(self, service, username, collection):
-            items = collection.search_items({"account": username, "service": service})
-            for item in items:
-                if hasattr(item, "unlock"):
-                    if item.is_locked() and item.unlock()[0]:
-                        raise keyring.errors.InitError("failed to unlock item")
-                return item.get_secret().decode("utf-8")
-
-        def get_password(self, service, username):
-            """Get password of the username for the service."""
-            collection = self.get_preferred_collection()
-            if hasattr(collection, "connection"):
-                with closing(collection.connection):
-                    return self.__get_password(service, username, collection)
-            else:
-                return self.__get_password(service, username, collection)
-
-    keyring.set_keyring(KeyringBackend())